requirements.txt
scikit-learn==1.3.1
pandas==2.1.1
pydantic==2.3.0
redis==5.0.0
pydantic-settings==2.0.3
celery==5.3.4
fastapi==0.103.1
uvicorn==0.23.2
google-api-python-client==2.101.0
google-auth-oauthlib==1.1.0